refactor: accounts views
Move user-account-related views and urls to `accounts.views` and `accounts.urls` respectively.
